+++ src/webpack-config-builder.js
@@ -651,11 +651,16 @@ export default class WebpackConfigBuilder {
 
     return {
       port: this.config.devServerPort,
-      contentBase: this.config.outputPath,
-      publicPath: '/',
-      compress: true,
-      writeToDisk: filePath => zipRegEx.test(filePath),
-      inline: false
+      magicHtml: false,
+      hot: false,
+      liveReload: false,
+      static: {
+        directory: this.config.outputPath,
+      },
+      client: false,
+      devMiddleware: {
+        writeToDisk: filePath => zipRegEx.test(filePath),
+      }
     };
   }
 
@@ -774,9 +779,9 @@ export default class WebpackConfigBuilder {
     buildConfigs.forEach((config, index) => {
       if (index === 0) {
         commonDevServerPort = config.devServer.port;
-        commonContentBase = config.devServer.contentBase;
+        commonContentBase = config.devServer.static.directory;
       } else {
-        commonContentBase = findStringSimilarities(commonContentBase, config.devServer.contentBase);
+        commonContentBase = findStringSimilarities(commonContentBase, config.devServer.static.directory);
       }
 
       if (index > 0) {
@@ -786,7 +791,7 @@ export default class WebpackConfigBuilder {
 
     let devServerConfig = buildConfigs[0].devServer;
     devServerConfig.port = commonDevServerPort;
-    devServerConfig.contentBase = commonContentBase;
+    devServerConfig.static.directory = commonContentBase;
 
     return buildConfigs;
   }